sys/handles: remove usage of docForward
{.docForward.} is limiting as it does not allow runnableExamples to be used.
Also with IC on the horizon Nim might remove the concept of forward declaration
altogether, so we migrate away from that too.

This commit convert the current implementation into {.dirty.} templates, which
is the currently recommended alternative. A perk of this is that we can finally
scope FDImpl to OS-specific implementation files.