sys/handles: remove isInvalidFD
The procedure was more or less a hack to avoid conversion warning. In true Nim
fashion, we fix this by introducing a special type of `InvalidFD` that is
hidden from the user.  In addition to this, we add `==`(AnyFD, typeof
InvalidFD) so that all `FD`-derived types might compare against `InvalidFD`.