sys/eventqueue: introduce a CPS event queue
Supports only Linux at the moment.

The eventqueue provides the magic `wait` which allow users to
specify what OS resource/event they are waiting for and allow
that time to be dispatched to an another task waiting to be run.

There is no tracking outside of CPS yet, and testing is very limited.

Currently the design put this eventqueue as a "full" queue that manages
every continuation, but the goal in the future is make it composable
so that users who want to build a dispatcher that dispatches to multiple
threads can do so.