sys(treewide): change IO semantic to oneshot
`sys` IO primitives was designed around reading a full file/pipe, which is
not something that usually happen and was an oversight.

Reimplemented the IO primitives to perform operations only once and not
try to fill the buffer anymore. This is less smart and much more
predictable.

For asynchronous IO, the primitives will now perform the operation, and
if that operation is not possible yet, wait until it is possible then
execute it.