sys/ioqueue(windows): introduce a persistence model
This feature would enable users to signal to the queue that an FD would
be producing a lot of events and they would like to avoid registration
overhead.

This is mandatory for anything IOCP on Windows, as IOCP requires
registration before any overlapped operation.

In the future this will be expanded to other OS, this is implemented
early for Windows so asynchronous operations can be performed
correctly, as such the semantics will change drastically in the future.