handles: change storage strategy and conform API
Since https://github.com/nim-lang/Nim/issues/19139 was closed as a
"won't fix", the "workaround" has to become permanent.

If we are to store a boolean next to the FD, it would double the
structure size, which is undesirable.

Instead, storage is now done via a lossless map so that zero-ed values
are invalid by default.

In this commit, Handle[T] API is changed to match other parts of the
project. In short:

* `get` -> `fd`
* `take` -> `takeFd`