sockets: prototype for ipv6
With the current prototype, I'm opting for keeping the transport protocol
address family agnostic. This means that variants have to be used to distinguish
between IPv4 and IPv6 addresses, but simplifies the typical usage of data transport
drastically by not requiring strict type disambiguation.

I might revisit making TCP4/TCP6 sockets a thing when a form of polymorphism exists
for the language.

Breaking changes:

- resolveIP4 -> resolveIP
- addressing APIs now return IPEndpoint, which could be either IP4 or IP6 endpoint