sockets: drop AI_ADDRCONFIG from getaddrinfo calls
This flag implementation is known to be troublesome:
https://fedoraproject.org/wiki/QA/Networking/NameResolution/ADDRCONFIG

Also, it would prevent users from explicitly requesting for IPv6
addresses on systems where *outbound* IPv6 is not available, which might
be a surprising behavior.

According to Fedora's documentation, glibc will sort the addresses so
that the best address is the first one, so this flag does not give us a
lot in terms of optimization there.