files: ignore cleanup exceptions on =destroy
In these situations there are only two reasons why cleaning up fails:

* The FD is invalid
* There is something wrong with the OS

The first case is already covered by Defects, which are not ignored,
and we can't do anything (nor can the caller) for the latter, so just
ignore those.

This is brought on by nimskull 0.1.0-dev.21268, which disallows raise
in hooks since that was undefined behavior.